chore: rename udt subcommand to sudt
