+++ src/subcommands/mod.rs
@@ -7,10 +7,10 @@ pub mod molecule;
 pub mod plugin;
 pub mod pubsub;
 pub mod rpc;
+pub mod sudt;
 #[cfg(unix)]
 pub mod tui;
 pub mod tx;
-pub mod udt;
 pub mod util;
 pub mod wallet;
 
@@ -26,8 +26,8 @@ pub use molecule::MoleculeSubCommand;
 pub use plugin::PluginSubCommand;
 pub use pubsub::PubSubCommand;
 pub use rpc::RpcSubCommand;
+pub use sudt::SudtSubCommand;
 pub use tx::TxSubCommand;
-pub use udt::{SudtSubCommand, UdtSubCommand};
 pub use util::UtilSubCommand;
 pub use wallet::{start_index_thread, LiveCells, TransferArgs, WalletSubCommand};
