+++ test/src/spec/udt/sudt.rs
@@ -22,7 +22,7 @@ impl Spec for SudtIssueToCheque {
         prepare(setup, path);
 
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:2000 --to-cheque-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:2000 --to-cheque-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR, ACCOUNT1_ADDR, cell_deps_path, owner_key_path
         ));
         log::info!("Issue 2000 SUDT to a cheque address which the sender is the sudt owner and the receiver is account 1:\n{}", output);
@@ -80,7 +80,7 @@ impl Spec for SudtIssueToCheque {
         );
 
         let output = setup.cli(&format!(
-            "udt transfer --owner {} --sender {} --udt-to {}:600 --to-acp-address --cell-deps {} --capacity-provider {} --privkey-path {} --privkey-path {}",
+            "sudt transfer --owner {} --sender {} --udt-to {}:600 --to-acp-address --cell-deps {} --capacity-provider {} --privkey-path {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             account2_acp_addr,
@@ -139,7 +139,7 @@ impl Spec for SudtIssueToAcp {
             0,
         );
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:300 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:300 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             cell_deps_path,
@@ -167,7 +167,7 @@ impl Spec for SudtIssueToAcp {
             account2_key_path.as_str(),
         );
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:200 --udt-to {}:400 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:200 --udt-to {}:400 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             account2_acp_addr,
@@ -231,7 +231,7 @@ impl Spec for SudtTransferToMultiAcp {
             account2_key_path.as_str(),
         );
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:300 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:300 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             cell_deps_path,
@@ -240,7 +240,7 @@ impl Spec for SudtTransferToMultiAcp {
         log::info!("Issue 300 SUDT to account 1's acp address:\n{}", output);
         setup.miner().generate_blocks(3);
         let output = setup.cli(&format!(
-            "udt transfer --owner {} --sender {} --udt-to {}:150 --udt-to {}:100 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt transfer --owner {} --sender {} --udt-to {}:150 --udt-to {}:100 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             owner_acp_addr,
@@ -305,7 +305,7 @@ impl Spec for SudtTransferToChequeForClaim {
         );
 
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:1100 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:1100 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             cell_deps_path,
@@ -326,7 +326,7 @@ impl Spec for SudtTransferToChequeForClaim {
         );
 
         let output = setup.cli(&format!(
-            "udt transfer --owner {} --sender {} --udt-to {}:500 --to-cheque-address --cell-deps {} --privkey-path {}",
+            "sudt transfer --owner {} --sender {} --udt-to {}:500 --to-cheque-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             ACCOUNT2_ADDR,
@@ -398,7 +398,7 @@ impl Spec for SudtTransferToChequeForWithdraw {
             account1_key_path.as_str(),
         );
         let output = setup.cli(&format!(
-            "udt issue --owner {} --udt-to {}:1100 --to-acp-address --cell-deps {} --privkey-path {}",
+            "sudt issue --owner {} --udt-to {}:1100 --to-acp-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             cell_deps_path,
@@ -419,7 +419,7 @@ impl Spec for SudtTransferToChequeForWithdraw {
         );
 
         let output = setup.cli(&format!(
-            "udt transfer --owner {} --sender {} --udt-to {}:500 --to-cheque-address --cell-deps {} --privkey-path {}",
+            "sudt transfer --owner {} --sender {} --udt-to {}:500 --to-cheque-address --cell-deps {} --privkey-path {}",
             OWNER_ADDR,
             account1_acp_addr,
             ACCOUNT2_ADDR,
